+++ docs/changelog.md
@@ -4,6 +4,8 @@
 
 - `borax.system` 模块
   - 新增 `rotate_filename` 文件名转换函数
+- `counters.serials` 模块
+  - 修正 `generate_serials` 异常提示错误
 
 ## v1.4.0 (20190519)